add implementation

+++ Src/ST-LIB_LOW/Supercaps/Supercaps.cpp
@@ -1,4 +1,4 @@
-#include "Supercaps.hpp"
+#include "ST-LIB_LOW/Supercaps/Supercaps.hpp"
 
 void Module::set_module_states(uint16_t module_state){
    is_active_cell_measuring = module_state & ACTIVE_CELL_MEASURE;
@@ -11,5 +11,9 @@ void Module::set_operating_mode(AcceptAutoTX accept_tx,StatusModule query_status
                         StatusCells query_status_cell,StatusDetail query_status_detail,
                         CellVoltages query_cell_voltages)
 {
-    operating_mode = accept_tx + query_status_module * 2 + query_status_cell * 4 + query_status_detail * 8 + query_cell_voltages * 16;
+    operating_mode = static_cast<uint8_t>(accept_tx) + 
+                     static_cast<uint8_t>(query_status_module) * 2 + 
+                     static_cast<uint8_t>(query_status_cell)   * 4 + 
+                     static_cast<uint8_t>(query_status_detail) * 8 + 
+                     static_cast<uint8_t>(query_cell_voltages) * 16;
 }
